pyproject.toml
[project]
name = "faq-bot"
version = "0.1.0"
description = "A chatbot for car insurance FAQ"
readme = "README.md"
requires-python = ">=3.7"

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0"
]